requirements.txt
flask
requests
python-dotenv
mcp
aiohttp
uvicorn
fastapi
anthropic
transformers
torch